+++ src/pages/blog/_components/ArticleCard.astro
@@ -2,7 +2,7 @@
 import { Icon } from "astro-icon/components";
 import type { CollectionEntry } from "astro:content";
 import Pill from "@components/Pills.svelte";
-import GitHubProfilePicture from "@components/GitHubProfilePicture.svelte";
+import ProfilePicture from "@components/ProfilePicture.svelte";
 import Link from "@components/Link.svelte";
 
 interface Props {
@@ -32,7 +32,7 @@ const formattedDatePosted = new Date(post.data.datePosted).toLocaleDateString("e
         <Pill list={[post.data.category]} />
         <div class="article-meta">
           <div class="article-author has-icon">
-            <GitHubProfilePicture username={post.data.author.github} size={64} wxh={24} />
+            <ProfilePicture username={post.data.author.github} size={64} wxh={24} />
             {post.data.author.name}
           </div>
           <p class="article-pubdate has-icon">
