+++ src/pages/community/_components/UserCard.astro
@@ -3,7 +3,7 @@ import type { ColorName } from "@catppuccin/palette";
 import maintainersWithoutAvatars from "@data/maintainersWithoutAvatars.json";
 
 import Link from "@components/Link.svelte";
-import GitHubProfilePicture from "@components/GitHubProfilePicture.svelte";
+import ProfilePicture from "@components/ProfilePicture.svelte";
 
 const isProd = import.meta.env.PROD;
 
@@ -21,7 +21,7 @@ const isPlaceholder = (maintainersWithoutAvatars as string[]).includes(username)
 
 <div class="card-user" style={`--user-color: var(--${color})`}>
   <Link {href} external>
-    <GitHubProfilePicture {username} size={256} wxh={128} />
+    <ProfilePicture {username} size={256} wxh={128} />
     <br />
     <span class=`user-name ${isPlaceholder ? "placeholder" : ""}`>{name}</span></Link
   >
