+++ src/pages/ports/_components/PortMaintainers.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import type { Port } from "@data/ports";
   import NoMaintainerIcon from "./NoMaintainerIcon.svelte";
-  import GitHubProfilePicture from "@components/GitHubProfilePicture.svelte";
+  import ProfilePicture from "@components/ProfilePicture.svelte";
 
   export let port: Port & { icon: string };
 </script>
@@ -9,7 +9,7 @@
 {#if port.repository["current-maintainers"].length > 0}
   <div class="current-maintainers" title="All users who currently maintain this port">
     {#each port.repository["current-maintainers"] as maintainer}
-      <GitHubProfilePicture username={maintainer.username} size={64} wxh={32} />
+      <ProfilePicture username={maintainer.username} size={64} wxh={32} />
     {/each}
   </div>
 {:else}
